conn: handle initial error packet correctly
If we haven't completed the hashshake the server will not be aware of our
capabilities and so its will packets behave as if we have none. This is necessary to
correcly parse an initial error packet which never contains an SQL State field even
if the client capabilities will eventually contain CLIENT_PROTOCOL_41.

https://dev.mysql.com/doc/dev/mysql-server/latest/page_protocol_connection_phase.html

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>